aliya_assistant/requirements.txt
Django==5.1.1
openai==1.40.0
python-dotenv==1.0.1
httpx==0.27.2
requests
bs4
